A bunch of fixes

+++ src/main/java/ru/javawebinar/topjava/repository/jdbc/JdbcUserRepository.java
@@ -9,25 +9,18 @@ import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import org.springframework.transaction.PlatformTransactionManager;
-import org.springframework.transaction.TransactionDefinition;
-import org.springframework.transaction.TransactionStatus;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.transaction.support.DefaultTransactionDefinition;
 import ru.javawebinar.topjava.model.Role;
 import ru.javawebinar.topjava.model.User;
 import ru.javawebinar.topjava.repository.UserRepository;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 import static java.sql.Types.INTEGER;
 import static java.sql.Types.VARCHAR;
-import static java.util.Collections.nCopies;
+import static java.util.EnumSet.noneOf;
 import static java.util.stream.Collectors.toList;
 import static ru.javawebinar.topjava.util.ValidationUtil.validateEntity;
 
@@ -41,8 +34,6 @@ public class JdbcUserRepository implements UserRepository {
 
     private final SimpleJdbcInsert insertUser;
 
-    private PlatformTransactionManager transactionManager;
-
     @Autowired
     public JdbcUserRepository(JdbcTemplate jdbcTemplate, NamedParameterJdbcTemplate namedParameterJdbcTemplate) {
         this.insertUser = new SimpleJdbcInsert(jdbcTemplate)
@@ -51,12 +42,6 @@ public class JdbcUserRepository implements UserRepository {
 
         this.jdbcTemplate = jdbcTemplate;
         this.namedParameterJdbcTemplate = namedParameterJdbcTemplate;
-
-    }
-
-    @Autowired
-    public void setTransactionManager(PlatformTransactionManager transactionManager) {
-        this.transactionManager = transactionManager;
     }
 
     @Override
@@ -64,8 +49,6 @@ public class JdbcUserRepository implements UserRepository {
     public User save(User user) {
         validateEntity(user);
         BeanPropertySqlParameterSource parameterSource = new BeanPropertySqlParameterSource(user);
-        TransactionDefinition definition = new DefaultTransactionDefinition();
-        TransactionStatus status = transactionManager.getTransaction(definition);
         if (user.isNew()) {
             Number newKey = insertUser.executeAndReturnKey(parameterSource);
             int userId = newKey.intValue();
@@ -81,66 +64,14 @@ public class JdbcUserRepository implements UserRepository {
             }
             updateRoles(user);
         }
-        transactionManager.commit(status);
         return user;
-    }
-
-    @Override
-    @Transactional
-    public boolean delete(int id) {
-        TransactionDefinition definition = new DefaultTransactionDefinition();
-        TransactionStatus status = transactionManager.getTransaction(definition);
-        int update = jdbcTemplate.update("DELETE FROM users WHERE id=?", id);
-        transactionManager.commit(status);
-        return update != 0;
-    }
-
-    @Override
-    public User get(int id) {
-        TransactionDefinition definition = new DefaultTransactionDefinition();
-        TransactionStatus status = transactionManager.getTransaction(definition);
-        List<User> query = jdbcTemplate.query("""
-                                SELECT *
-                                FROM users
-                                JOIN user_roles ur ON users.id = ur.user_id
-                                WHERE users.id=?
-                """, new UserWithRolesExtractor(), id);
-        transactionManager.commit(status);
-        return DataAccessUtils.singleResult(query);
-    }
-
-    @Override
-    public User getByEmail(String email) {
-        TransactionDefinition definition = new DefaultTransactionDefinition();
-        TransactionStatus status = transactionManager.getTransaction(definition);
-        List<User> users = jdbcTemplate.query("""
-                SELECT * 
-                FROM users 
-                JOIN user_roles ur ON users.id = ur.user_id
-                WHERE email=?
-                """, new UserWithRolesExtractor(), email);
-        transactionManager.commit(status);
-        return DataAccessUtils.singleResult(users);
-    }
-
-    @Override
-    public List<User> getAll() {
-        TransactionDefinition definition = new DefaultTransactionDefinition();
-        TransactionStatus status = transactionManager.getTransaction(definition);
-        List<User> query = jdbcTemplate.query("""
-                SELECT * 
-                FROM users
-                JOIN user_roles ur ON users.id = ur.user_id 
-                ORDER BY name, email""", new UserWithRolesExtractor());
-        transactionManager.commit(status);
-        return query;
     }
 
     private void insertRoles(final User user) {
         int[] argTypes = {VARCHAR, INTEGER};
         List<Object[]> batchArgs = user.getRoles()
                 .stream()
-                .map(Role::toString)
+                .map(Role::name)
                 .map(str -> new Object[]{str, user.getId()})
                 .collect(toList());
         jdbcTemplate.batchUpdate("""
@@ -150,26 +81,56 @@ public class JdbcUserRepository implements UserRepository {
     }
 
     private void updateRoles(final User user) {
-        int size = user.getRoles().size();
-        List<Object[]> batchArgs = nCopies(size, new Object[]{user.getId()});
-        jdbcTemplate.batchUpdate("""
-                DELETE FROM user_roles
-                WHERE user_id = ?
-                """, batchArgs);
+        jdbcTemplate.update("DELETE FROM user_roles WHERE user_id=?", user.getId());
         insertRoles(user);
     }
 
-    private static class UserWithRolesExtractor implements ResultSetExtractor<ArrayList<User>> {
+    @Override
+    @Transactional
+    public boolean delete(int id) {
+        return jdbcTemplate.update("DELETE FROM users WHERE id=?", id) != 0;
+    }
+
+    @Override
+    public User get(int id) {
+        List<User> query = jdbcTemplate.query("""
+                SELECT *
+                FROM users
+                LEFT JOIN user_roles ur ON users.id = ur.user_id
+                WHERE users.id=?
+                """, new UserWithRolesExtractor(), id);
+        return DataAccessUtils.singleResult(query);
+    }
+
+    @Override
+    public User getByEmail(String email) {
+        List<User> users = jdbcTemplate.query("""
+                SELECT * 
+                FROM users 
+                LEFT JOIN user_roles ur ON users.id = ur.user_id
+                WHERE email=?
+                """, new UserWithRolesExtractor(), email);
+        return DataAccessUtils.singleResult(users);
+    }
+
+    @Override
+    public List<User> getAll() {
+        return jdbcTemplate.query("""
+                SELECT * 
+                FROM users
+                LEFT JOIN user_roles ur ON users.id = ur.user_id 
+                ORDER BY name, email""", new UserWithRolesExtractor());
+    }
+
+    private static class UserWithRolesExtractor implements ResultSetExtractor<List<User>> {
 
         @Override
-        public ArrayList<User> extractData(ResultSet rs) throws SQLException, DataAccessException {
-            Map<Integer, User> map = new HashMap<>();
-            User user;
+        public List<User> extractData(ResultSet rs) throws SQLException, DataAccessException {
+            Map<Integer, User> map = new LinkedHashMap<>();
             while (rs.next()) {
                 int id = rs.getInt("id");
-                user = map.get(id);
-                if (user == null) {
-                    user = new User();
+                User user = map.computeIfAbsent(id, integer -> new User());
+                if (user.isNew()) {
                     user.setId(id);
                     user.setName(rs.getString("name"));
                     user.setPassword(rs.getString("password"));
@@ -177,12 +138,12 @@ public class JdbcUserRepository implements UserRepository {
                     user.setEnabled(rs.getBoolean("enabled"));
                     user.setCaloriesPerDay(rs.getInt("calories_per_day"));
                     user.setRegistered(rs.getDate("registered"));
-                    user.setRoles(new ArrayList<>());
-                    map.put(id, user);
+                    user.setRoles(noneOf(Role.class));
                 }
 
-                String role = rs.getString("role");
-                user.getRoles().add(Role.valueOf(role));
+                Optional.ofNullable(rs.getString("role"))
+                        .map(Role::valueOf)
+                        .ifPresent(user::addRole);
             }
             return new ArrayList<>(map.values());
         }
